+++ shared/src/commonMain/kotlin/org/hyperskill/app/core/injection/StateRepositoriesComponentImpl.kt
@@ -52,6 +52,7 @@ class StateRepositoriesComponentImpl(appGraph: AppGraph) : StateRepositoriesComp
         )
     }
 
-    override val interviewStepsStateRepository: InterviewStepsStateRepository =
+    override val interviewStepsStateRepository: InterviewStepsStateRepository by lazy {
         InterviewStepsStateRepositoryImpl()
+    }
 }
