+++ shared/src/commonMain/kotlin/org/hyperskill/app/interview_preparation/domain/analytic/InterviewPreparationWidgetClickedRetryContentLoadingHyperskillAnalyticsEvent.kt
@@ -1,25 +1,28 @@
-package org.hyperskill.app.interview_preparation.domain.analytics
+package org.hyperskill.app.interview_preparation.domain.analytic
 
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticAction
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticEvent
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticPart
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticRoute
+import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticTarget
 
 /**
- * Represents a click analytic event of the error state placeholder retry button.
+ * Represents an analytic event for clicking on a retry button in the interview preparation widget.
  *
  * JSON payload:
  * ```
  * {
  *     "route": "/home",
  *     "action": "click",
- *     "part": "interview_preparation_widget"
+ *     "part": "interview_preparation_widget",
+ *     "target": "retry"
  * }
  * ```
  * @see HyperskillAnalyticEvent
  */
-object InterviewPreparationWidgetClickedHyperskillAnalyticsEvent : HyperskillAnalyticEvent(
+object InterviewPreparationWidgetClickedRetryContentLoadingHyperskillAnalyticsEvent : HyperskillAnalyticEvent(
     HyperskillAnalyticRoute.Home(),
     HyperskillAnalyticAction.CLICK,
-    HyperskillAnalyticPart.INTERVIEW_PREPARATION_WIDGET
+    HyperskillAnalyticPart.INTERVIEW_PREPARATION_WIDGET,
+    HyperskillAnalyticTarget.RETRY
 )
