+++ shared/src/commonMain/kotlin/org/hyperskill/app/interview_preparation/presentation/InterviewPreparationWidgetActionDispatcher.kt
@@ -39,6 +39,7 @@ class InterviewPreparationWidgetActionDispatcher(
             }
             .launchIn(actionScope)
     }
+
     override suspend fun doSuspendableAction(action: Action) {
         when (action) {
             InternalAction.FetchInterviewSteps -> {
