+++ shared/src/commonMain/kotlin/org/hyperskill/app/interview_preparation/presentation/InterviewPreparationWidgetReducer.kt
@@ -1,8 +1,8 @@
 package org.hyperskill.app.interview_preparation.presentation
 
-import org.hyperskill.app.interview_preparation.domain.analytics.InterviewPreparationWidgetClickedHyperskillAnalyticsEvent
-import org.hyperskill.app.interview_preparation.domain.analytics.InterviewPreparationWidgetClickedRetryContentLoadingHyperskillAnalyticsEvent
-import org.hyperskill.app.interview_preparation.domain.analytics.InterviewPreparationWidgetViewedHyperskillAnalyticsEvent
+import org.hyperskill.app.interview_preparation.domain.analytic.InterviewPreparationWidgetClickedHyperskillAnalyticsEvent
+import org.hyperskill.app.interview_preparation.domain.analytic.InterviewPreparationWidgetClickedRetryContentLoadingHyperskillAnalyticsEvent
+import org.hyperskill.app.interview_preparation.domain.analytic.InterviewPreparationWidgetViewedHyperskillAnalyticsEvent
 import org.hyperskill.app.interview_preparation.presentation.InterviewPreparationWidgetFeature.Action
 import org.hyperskill.app.interview_preparation.presentation.InterviewPreparationWidgetFeature.InternalAction
 import org.hyperskill.app.interview_preparation.presentation.InterviewPreparationWidgetFeature.InternalMessage
