+++ shared/src/commonMain/kotlin/org/hyperskill/app/interview_preparation/view/model/InterviewPreparationWidgetViewState.kt
@@ -7,7 +7,7 @@ sealed interface InterviewPreparationWidgetViewState {
     object Empty : InterviewPreparationWidgetViewState
 
     data class Content(
-        val stepsAmount: Int,
+        val stepsCount: Int,
         val description: String
     ) : InterviewPreparationWidgetViewState
 }
